requirements.txt
fastapi
uvicorn[standard]
requests
transformers
torch
pillow
python-multipart
pytest
